merge: (#507) Exception 파일 병합
