--- dms-core/src/main/kotlin/team/aliens/dms/domain/file/exception/FileIOInterruptedException.kt
@@ -1,8 +1,0 @@
-package team.aliens.dms.domain.file.exception
-
-import team.aliens.dms.common.error.DmsException
-import team.aliens.dms.domain.file.exception.error.FileErrorCode
-
-object FileIOInterruptedException : DmsException(
-    FileErrorCode.IO_INTERRUPTED
-)
