+++ dms-core/src/main/kotlin/team/aliens/dms/domain/manager/exception/ManagerErrorCode.kt
@@ -1,18 +1,18 @@
-package team.aliens.dms.domain.remain.exception.error
+package team.aliens.dms.domain.manager.exception
 
 import team.aliens.dms.common.error.ErrorProperty
 import team.aliens.dms.common.error.ErrorStatus
 
-enum class RemainOptionErrorCode(
+enum class ManagerErrorCode(
     private val status: Int,
     private val message: String,
     private val sequence: Int
 ) : ErrorProperty {
 
-    REMAIN_OPTION_NOT_FOUND(ErrorStatus.NOT_FOUND, "Remain Option Not Found", 2)
+    MANAGER_INFO_NOT_MATCHED(ErrorStatus.UNAUTHORIZED, "Manager Info Not Matched", 1)
     ;
 
     override fun status(): Int = status
     override fun message(): String = message
-    override fun code(): String = "REMAIN-$status-$sequence"
+    override fun code(): String = "MANAGER-$status-$sequence"
 }
