--- dms-core/src/main/kotlin/team/aliens/dms/domain/manager/exception/ManagerNotFoundException.kt
@@ -1,8 +1,0 @@
-package team.aliens.dms.domain.manager.exception
-
-import team.aliens.dms.common.error.DmsException
-import team.aliens.dms.domain.manager.exception.error.ManagerErrorCode
-
-object ManagerNotFoundException : DmsException(
-    ManagerErrorCode.MANAGER_NOT_FOUND
-)
