+++ dms-core/src/main/kotlin/team/aliens/dms/domain/notice/exception/NoticeExceptions.kt
@@ -1,7 +1,10 @@
 package team.aliens.dms.domain.notice.exception
 
 import team.aliens.dms.common.error.DmsException
-import team.aliens.dms.domain.notice.exception.error.NoticeErrorCode
+
+object NoticeNotFoundException : DmsException(
+    NoticeErrorCode.NOTICE_NOT_FOUND
+)
 
 object IsNotWriterException : DmsException(
     NoticeErrorCode.IS_NOT_WRITER
