--- dms-core/src/main/kotlin/team/aliens/dms/domain/notice/exception/NoticeNotFoundException.kt
@@ -1,8 +1,0 @@
-package team.aliens.dms.domain.notice.exception
-
-import team.aliens.dms.common.error.DmsException
-import team.aliens.dms.domain.notice.exception.error.NoticeErrorCode
-
-object NoticeNotFoundException : DmsException(
-    NoticeErrorCode.NOTICE_NOT_FOUND
-)
