--- dms-core/src/main/kotlin/team/aliens/dms/domain/point/exception/InvalidPointFilterRangeException.kt
@@ -1,8 +1,0 @@
-package team.aliens.dms.domain.point.exception
-
-import team.aliens.dms.common.error.DmsException
-import team.aliens.dms.domain.point.exception.error.PointHistoryErrorCode
-
-object InvalidPointFilterRangeException : DmsException(
-    PointHistoryErrorCode.INVALID_POINT_FILTER_RANGE
-)
