--- dms-core/src/main/kotlin/team/aliens/dms/domain/point/exception/PointHistoryNotFoundException.kt
@@ -1,8 +1,0 @@
-package team.aliens.dms.domain.point.exception
-
-import team.aliens.dms.common.error.DmsException
-import team.aliens.dms.domain.point.exception.error.PointHistoryErrorCode
-
-object PointHistoryNotFoundException : DmsException(
-    PointHistoryErrorCode.POINT_HISTORY_NOT_FOUND
-)
