--- dms-core/src/main/kotlin/team/aliens/dms/domain/point/exception/PointOptionNotFoundException.kt
@@ -1,8 +1,0 @@
-package team.aliens.dms.domain.point.exception
-
-import team.aliens.dms.common.error.DmsException
-import team.aliens.dms.domain.point.exception.error.PointOptionErrorCode
-
-object PointOptionNotFoundException : DmsException(
-    PointOptionErrorCode.POINT_OPTION_NOT_FOUND
-)
