--- dms-core/src/main/kotlin/team/aliens/dms/domain/point/exception/PointOptionSchoolMismatchException.kt
@@ -1,8 +1,0 @@
-package team.aliens.dms.domain.point.exception
-
-import team.aliens.dms.common.error.DmsException
-import team.aliens.dms.domain.point.exception.error.PointOptionErrorCode
-
-object PointOptionSchoolMismatchException : DmsException(
-    PointOptionErrorCode.POINT_OPTION_SCHOOL_MISMATCH
-)
