--- dms-core/src/main/kotlin/team/aliens/dms/domain/point/exception/error/PointHistoryErrorCode.kt
@@ -1,22 +1,0 @@
-package team.aliens.dms.domain.point.exception.error
-
-import team.aliens.dms.common.error.ErrorProperty
-import team.aliens.dms.common.error.ErrorStatus
-
-enum class PointHistoryErrorCode(
-    private val status: Int,
-    private val message: String,
-    private val sequence: Int
-) : ErrorProperty {
-
-    POINT_HISTORY_NOT_FOUND(ErrorStatus.NOT_FOUND, "Point History Not Found", 1),
-
-    INVALID_POINT_FILTER_RANGE(ErrorStatus.BAD_REQUEST, "Invalid Point Filter Range", 2),
-
-    POINT_HISTORY_CAN_NOT_CANCEL(ErrorStatus.BAD_REQUEST, "Point History Can Not Cancel", 2)
-    ;
-
-    override fun status(): Int = status
-    override fun message(): String = message
-    override fun code(): String = "POINT-$status-$sequence"
-}
