--- dms-core/src/main/kotlin/team/aliens/dms/domain/remain/exception/RemainAvailableTimeNotFoundException.kt
@@ -1,8 +1,0 @@
-package team.aliens.dms.domain.remain.exception
-
-import team.aliens.dms.common.error.DmsException
-import team.aliens.dms.domain.remain.exception.error.RemainAvailableTimeErrorCode
-
-object RemainAvailableTimeNotFoundException : DmsException(
-    RemainAvailableTimeErrorCode.REMAIN_AVAILABLE_TIME_NOT_FOUND
-)
