+++ dms-core/src/main/kotlin/team/aliens/dms/domain/remain/exception/RemainErrorCode.kt
@@ -1,18 +1,21 @@
-package team.aliens.dms.domain.remain.exception.error
+package team.aliens.dms.domain.remain.exception
 
 import team.aliens.dms.common.error.ErrorProperty
 import team.aliens.dms.common.error.ErrorStatus
 
-enum class RemainAvailableTimeErrorCode(
+enum class RemainErrorCode(
     private val status: Int,
     private val message: String,
     private val sequence: Int
 ) : ErrorProperty {
 
     REMAIN_CAN_NOT_APPLIED(ErrorStatus.FORBIDDEN, "Remain Can Not Apply", 1),
-    REMAIN_AVAILABLE_TIME_CAN_NOT_ACCESS(ErrorStatus.FORBIDDEN, "Remain Available Time Can Not Access", 2),
+    REMAIN__CAN_NOT_ACCESS(ErrorStatus.FORBIDDEN, "Remain Available Time Can Not Access", 2),
 
-    REMAIN_AVAILABLE_TIME_NOT_FOUND(ErrorStatus.NOT_FOUND, "Remain Available Time Not Found", 1),
+    REMAIN__NOT_FOUND(ErrorStatus.NOT_FOUND, "Remain Available Time Not Found", 1),
+    REMAIN_OPTION_NOT_FOUND(ErrorStatus.NOT_FOUND, "Remain Option Not Found", 2),
+
+    REMAIN_STATUS_NOT_FOUND(ErrorStatus.NOT_FOUND, "Remain Status Not Found", 3)
     ;
 
     override fun status() = status
