--- dms-core/src/main/kotlin/team/aliens/dms/domain/remain/exception/RemainOptionNotFoundException.kt
@@ -1,8 +1,0 @@
-package team.aliens.dms.domain.remain.exception
-
-import team.aliens.dms.common.error.DmsException
-import team.aliens.dms.domain.remain.exception.error.RemainOptionErrorCode
-
-object RemainOptionNotFoundException : DmsException(
-    RemainOptionErrorCode.REMAIN_OPTION_NOT_FOUND
-)
