--- dms-core/src/main/kotlin/team/aliens/dms/domain/remain/exception/RemainStatusNotFound.kt
@@ -1,8 +1,0 @@
-package team.aliens.dms.domain.remain.exception
-
-import team.aliens.dms.common.error.DmsException
-import team.aliens.dms.domain.remain.exception.error.RemainStatusErrorCode
-
-object RemainStatusNotFound : DmsException(
-    RemainStatusErrorCode.REMAIN_STATUS_NOT_FOUND
-)
