--- dms-core/src/main/kotlin/team/aliens/dms/domain/school/exception/SchoolCodeMismatchException.kt
@@ -1,8 +1,0 @@
-package team.aliens.dms.domain.school.exception
-
-import team.aliens.dms.common.error.DmsException
-import team.aliens.dms.domain.school.exception.error.SchoolErrorCode
-
-object SchoolCodeMismatchException : DmsException(
-    SchoolErrorCode.SCHOOL_CODE_MISMATCH
-)
