--- dms-core/src/main/kotlin/team/aliens/dms/domain/student/exception/StudentAlreadyExistsException.kt
@@ -1,8 +1,0 @@
-package team.aliens.dms.domain.student.exception
-
-import team.aliens.dms.common.error.DmsException
-import team.aliens.dms.domain.student.exception.error.StudentErrorCode
-
-object StudentAlreadyExistsException : DmsException(
-    StudentErrorCode.STUDENT_ALREADY_EXISTS
-)
