--- dms-core/src/main/kotlin/team/aliens/dms/domain/student/exception/VerifiedStudentNotFoundException.kt
@@ -1,8 +1,0 @@
-package team.aliens.dms.domain.student.exception
-
-import team.aliens.dms.common.error.DmsException
-import team.aliens.dms.domain.student.exception.error.VerifiedStudentErrorCode
-
-object VerifiedStudentNotFoundException : DmsException(
-    VerifiedStudentErrorCode.VERIFIED_STUDENT_NOT_FOUND
-)
