+++ dms-infrastructure/src/main/kotlin/team/aliens/dms/global/error/GlobalErrorCode.kt
@@ -15,9 +15,9 @@ enum class GlobalErrorCode(
     INVALID_FILE(ErrorStatus.BAD_REQUEST, "Invalid File", 4),
     BAD_EXCEL_FORMAT(ErrorStatus.BAD_REQUEST, "%s행 등 %s개 행의 데이터 형식이 잘못되었습니다.", 5),
 
-    EXTENSION_MISMATCH(401, "File Extension Mismatch", 1),
+    EXTENSION_MISMATCH(ErrorStatus.UNAUTHORIZED, "File Extension Mismatch", 1),
 
-    INTERNAL_SERVER_ERROR(500, "Internal Server Error", 1)
+    INTERNAL_SERVER_ERROR(ErrorStatus.INTERNAL_SERVER_ERROR, "Internal Server Error", 1)
     ;
 
     override fun status(): Int = status
